Implementing collision primitive importing

+++ blender/source/register/operators/mesh_info_operators.py
@@ -12,6 +12,11 @@ from .base import (
 from ...exceptions import HEIOUserException, HEIODevException
 from ...utility import attribute_utils
 
+NO_ATTRIB = {
+    'COLLISION_CONVEXFLAGS',
+    'COLLISION_PRIMITIVES',
+    'COLLISION_PRIMITIVEFLAGS'
+}
 
 class BaseMeshInfoOperator(HEIOBaseOperator):
     bl_options = {'UNDO'}
@@ -25,6 +30,8 @@ class BaseMeshInfoOperator(HEIOBaseOperator):
             ('COLLISION_TYPES', "", ""),
             ('COLLISION_FLAGS', "", ""),
             ('COLLISION_CONVEXFLAGS', "", ""),
+            ('COLLISION_PRIMITIVES', "", ""),
+            ('COLLISION_PRIMITIVEFLAGS', "", ""),
             ('ERROR', "", "")
         ),
         default='ERROR',
@@ -63,6 +70,15 @@ class BaseMeshInfoOperator(HEIOBaseOperator):
                 raise HEIOUserException("No active mesh collision layer!")
             return layer.convex_flags
 
+        elif self.type == 'COLLISION_PRIMITIVES':
+            return collision_info.primitives
+
+        elif self.type == 'COLLISION_PRIMITIVEFLAGS':
+            primitive = collision_info.primitives.active_element
+            if primitive is None:
+                raise HEIOUserException("No active mesh collision primitive!")
+            return primitive.surface_flags
+
         else:  # ERROR
             raise HEIOUserException("Invalid call!")
 
@@ -73,6 +89,9 @@ class HEIO_OT_MeshInfo_Initialize(BaseMeshInfoOperator):
     bl_description = "Set up the mesh to use certain mesh info"
 
     def _execute(self, context):
+        if self.type in NO_ATTRIB:
+            raise HEIODevException("Invalid type")
+
         self.get_list(context).initialize()
         return {'FINISHED'}
 
@@ -83,6 +102,9 @@ class HEIO_OT_MeshInfo_Delete(BaseMeshInfoOperator):
     bl_description = "Delete mesh info from the mesh"
 
     def _execute(self, context):
+        if self.type in NO_ATTRIB:
+            raise HEIODevException("Invalid type")
+
         self.get_list(context).delete()
         return {'FINISHED'}
 
@@ -96,7 +118,7 @@ class MeshInfoEditOperator(BaseMeshInfoOperator):
     def _execute(self, context):
         list = self.get_list(context)
 
-        if self.type == 'COLLISION_CONVEXFLAGS':
+        if self.type in NO_ATTRIB:
             raise HEIODevException("Invalid type")
 
         if not list.initialized:
@@ -195,14 +217,17 @@ class MeshListOperator(BaseMeshInfoOperator):
     def _execute(self, context):
         list = self.get_list(context)
 
-        if self.type != 'COLLISION_CONVEXFLAGS' and not list.initialized:
-            raise HEIOUserException("Not initialized!")
+        if self.type not in NO_ATTRIB :
+            if not list.initialized:
+                raise HEIOUserException("Not initialized!")
 
-        self.attribute_utility = attribute_utils.AttributeUtility(
-            context,
-            context.active_object.data,
-            list.attribute_name
-        )
+            self.attribute_utility = attribute_utils.AttributeUtility(
+                context,
+                context.active_object.data,
+                list.attribute_name
+            )
+        else:
+            self.attribute_utility = None
 
         self.list_execute(context, list)
         return {'FINISHED'}
@@ -220,7 +245,7 @@ class HEIO_OT_MeshInfo_Remove(MeshListOperator, ListRemove):
     bl_description = "Removes the selected mesh info item from the mesh"
 
     def list_execute(self, context: bpy.types.Context, target_list):
-        if self.type != 'COLLISION_CONVEXFLAGS':
+        if self.attribute_utility is not None:
             if context.mode == "EDIT_MESH":
                 raise HEIOUserException(
                     "Cannot remove mesh info while in edit mode!")
@@ -234,6 +259,9 @@ class HEIO_OT_MeshInfo_Remove(MeshListOperator, ListRemove):
 
         removed_index = super().list_execute(context, target_list)
 
+        if self.attribute_utility is None:
+            return
+
         if self.type == 'MESH_LAYERS':
             if removed_index == 3:
                 self.attribute_utility.change_int_values(removed_index, 0)
@@ -244,7 +272,7 @@ class HEIO_OT_MeshInfo_Remove(MeshListOperator, ListRemove):
         elif self.type == 'COLLISION_FLAGS':
             self.attribute_utility.rightshift_int_flags(removed_index, 1)
 
-        elif self.type != 'COLLISION_CONVEXFLAGS':
+        else:
             self.attribute_utility.decrease_int_values(max(removed_index, 1))
 
 
@@ -254,22 +282,23 @@ class HEIO_OT_MeshInfo_Move(MeshListOperator, ListMove):
     bl_description = "Moves the selected mesh info item in the mesh"
 
     def list_execute(self, context, target_list):
-        if self.type != 'COLLISION_CONVEXFLAGS' and context.mode == "EDIT_MESH":
-            raise HEIOUserException(
-                "Cannot move mesh info while in edit mode!")
-
-        if self.type == 'MESH_LAYERS':
-            if target_list.active_index < 3:
+        if self.attribute_utility is not None:
+            if context.mode == "EDIT_MESH":
                 raise HEIOUserException(
-                    f"The {target_list.active_element.name} layer cannot be moved!")
+                    "Cannot move mesh info while in edit mode!")
 
-            if target_list.active_index == 3 and self.direction == 'UP':
-                raise HEIOUserException(
-                    f"The {target_list.active_element.name} layer cannot be moved up here!")
+            if self.type == 'MESH_LAYERS':
+                if target_list.active_index < 3:
+                    raise HEIOUserException(
+                        f"The {target_list.active_element.name} layer cannot be moved!")
+
+                if target_list.active_index == 3 and self.direction == 'UP':
+                    raise HEIOUserException(
+                        f"The {target_list.active_element.name} layer cannot be moved up here!")
 
         old_index, new_index = super().list_execute(context, target_list)
 
-        if old_index is None:
+        if old_index is None or self.attribute_utility is not None:
             return
 
         if self.type == 'COLLISION_FLAGS':
